Lab6 makefile complete
